requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
httpx>=0.25.0
pydantic>=2.4.2
langchain-openai>=0.0.4
langgraph>=0.0.20
openai>=1.3.0
matplotlib>=3.8.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.25.0
pillow>=10.0.0
python-multipart>=0.0.6
requests>=2.31.0
PyMuPDF>=1.22.3
fitz>=0.0.1
PyPDF2>=3.0.0